pyproject.toml
[project]
name = "tracklab"
version = "1.3.20"
description = "A Modular End-to-End Tracking Framework for Research and Development"
readme = "README.md"
authors = [
    { name = "Victor Joos", email = "victorjoosdtb@gmail.com" },
    { name = "Baptiste Standaert", email = "baptiste.standaert@uclouvain.be"},
    { name = "Vladimir Somers", email = "vladimir.somers@uclouvain.be"},
]

requires-python = ">=3.9"
dependencies = [
    "hydra-core>=1.3",
    "lightning>=2.0",
    "pytorch_lightning>=2.0",
    "numpy>=1.26.4",
    "ultralytics>=8.0.100",
    "filterpy>=1.4.5",
    "torch>=1.13.1",
    "torchvision>=0.14.1",
    "timm>=0.9.12",
    "soccernet>=0.1.53",
    "yt-dlp>=2023.12.30",
    "gdown>=4.7.1",
    "pandas>=2.1.0",
    "matplotlib>=3.9.0",
    "scikit-learn>=1.5.1",
    "rich",
    "tabulate",
    "sn-trackeval>=0.3.0",
    "lap",
    "distinctipy",
    "rtmlib",
    "huggingface-hub",
    "opencv-python",
]

[project.optional-dependencies]
# Always install mmcv with mim install mmcv==2.2.0 when installing this extra
#openmmlab = [
#    "mmdet>=3.1.0",
#    "mmengine>=0.10.1",
#    "openmim>=0.3.9",
#    "torch<=2.1.2",
#    "torchvision<=0.16.2",
#    "mmpose>=1.2.0",
#]

transformers = [
    "transformers",
]

yolox = [
    "pixeltable-yolox"
]


[dependency-groups]
dev = [
    "tracklab[transformers,yolox]",
    "twine>=6.1.0",
    "huggingface_hub[cli]",
]

[project.scripts]
tracklab = "tracklab.main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".", "plugins/track"]
include = ["tracklab*", "bot_sort*", "bpbreid_strong_sort*", "byte_track*", "deep_oc_sort*", "oc_sort*", "strong_sort*", "hydra_plugins.*"]
exclude = []

[tool.setuptools.package-data]
"*" = ["*.txt", "*.rst", "*.yaml"]

[tool.uv]
# Ensure that the setuptools v60.0.0 is used whenever a package has a build dependency
# on setuptools.
no-build-isolation-package = ["chumpy"]
#find-links = ["https://download.openmmlab.com/mmcv/dist/cu121/torch2.1/index.html"]

#[[tool.uv.index]]
#name = "pytorch"
#url = "https://download.pytorch.org/whl/cu121"
#explicit = true
#
#[tool.uv.sources]
#torch = { index = "pytorch" }
#torchvision = {index = "pytorch"}
#conflicts = [
#    [
#      { extra = "openmmlab" },
#      { extra = "yolox" },
#    ],
#]